Proof-of-concept directory listing feature
This feature allows, for example, when someone visits "/var/www/static/subdirhere", to list all the contents of that directory. Note that this if the user enters the root URI as defined in the config (/var/www/static default) it will automatically fetch the index.html instead of the directory's content. This feature is still on the works.