Improve code reusability and stability improvements
This commit makes a few changes to how requests work.  and  are removed and replaced with a FinalResponse struct that implements the new ResponseFactory trait. Additionally, a few unwraps have been removed and replaced with manual error handling, and some reorganization was done to make the file names for each code utility more meaningful.